Expose all jack ports when launching the application
Due to a limitation in gstreamer's jackaudiosink, jack ports are only
created when first needed.

We here create an empty temporary wav file and play it, so the ports
will be available for connecting them immediately.

The proper solution would be an extension to jackaudiosink that allows
to expose the ports upon request.